requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
asyncio==3.4.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
croniter==5.0.1
discord-py-interactions==5.13.2
discord-typings==0.9.0
discord.py==2.4.0
emoji==2.14.0
frozenlist==1.4.1
idna==3.7
logging==0.4.9.6
multidict==6.0.5
pyfiglet==1.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
tls-client==1.0.1
tomli==2.1.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.9.4
